+++ internal/checks/base.go
@@ -33,6 +33,7 @@ var (
 		RuleDependencyCheckName,
 		RuleDuplicateCheckName,
 		RuleForCheckName,
+		RuleNameCheckName,
 		LabelCheckName,
 		RuleLinkCheckName,
 		RejectCheckName,