+++ internal/checks/rule_name_test.go
@@ -1,0 +1,77 @@
+package checks_test
+
+import (
+	"testing"
+
+	"github.com/cloudflare/pint/internal/checks"
+	"github.com/cloudflare/pint/internal/parser"
+	"github.com/cloudflare/pint/internal/promapi"
+)
+
+func TestRuleName(t *testing.T) {
+	testCases := []checkTest{
+		{
+			description: "doesn't ignore rules with syntax errors",
+			content:     "- record: foo\n  expr: sum(foo) without(\n",
+			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
+				return checks.NewRuleNameCheck(checks.MustTemplatedRegexp("total:.+"), "some text", checks.Warning)
+			},
+			prometheus: noProm,
+			problems: func(_ string) []checks.Problem {
+				return []checks.Problem{
+					{
+						Lines: parser.LineRange{
+							First: 1,
+							Last:  1,
+						},
+						Reporter: checks.RuleNameCheckName,
+						Text:     "recording rule name must match `^total:.+$`.",
+						Details:  "Rule comment: some text",
+						Severity: checks.Warning,
+					},
+				}
+			},
+		},
+		{
+			description: "doesn't ignore rules with syntax errors",
+			content:     "- record: total:foo\n  expr: sum(foo) without(\n",
+			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
+				return checks.NewRuleNameCheck(checks.MustTemplatedRegexp("total:.+"), "some text", checks.Warning)
+			},
+			prometheus: noProm,
+			problems:   noProblems,
+		},
+		{
+			description: "doesn't ignore rules with syntax errors",
+			content:     "- alert: foo\n  expr: sum(foo) without(\n",
+			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
+				return checks.NewRuleNameCheck(checks.MustTemplatedRegexp("total:.+"), "some text", checks.Warning)
+			},
+			prometheus: noProm,
+			problems: func(_ string) []checks.Problem {
+				return []checks.Problem{
+					{
+						Lines: parser.LineRange{
+							First: 1,
+							Last:  1,
+						},
+						Reporter: checks.RuleNameCheckName,
+						Text:     "alerting rule name must match `^total:.+$`.",
+						Details:  "Rule comment: some text",
+						Severity: checks.Warning,
+					},
+				}
+			},
+		},
+		{
+			description: "doesn't ignore rules with syntax errors",
+			content:     "- alert: total:foo\n  expr: sum(foo) without(\n",
+			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
+				return checks.NewRuleNameCheck(checks.MustTemplatedRegexp("total:.+"), "some text", checks.Warning)
+			},
+			prometheus: noProm,
+			problems:   noProblems,
+		},
+	}
+	runTests(t, testCases)
+}