+++ internal/config/config_test.go
@@ -1547,6 +1547,31 @@ rule {
 			},
 		},
 		{
+			title: "name",
+			config: `
+rule {
+  name "total:.+" {}
+}
+`,
+			entry: discovery.Entry{
+				State: discovery.Modified,
+				Path: discovery.Path{
+					Name:          "rules.yml",
+					SymlinkTarget: "rules.yml",
+				},
+				Rule: newRule(t, "- record: foo\n  expr: sum(foo)\n"),
+			},
+			checks: []string{
+				checks.SyntaxCheckName,
+				checks.AlertForCheckName,
+				checks.ComparisonCheckName,
+				checks.TemplateCheckName,
+				checks.FragileCheckName,
+				checks.RegexpCheckName,
+				checks.RuleNameCheckName + "(^total:.+$)",
+			},
+		},
+		{
 			title: "two prometheus servers / disable checks via file/disable comment",
 			config: `
 prometheus "prom1" {
@@ -2221,6 +2246,20 @@ func TestConfigErrors(t *testing.T) {
   }
 }`,
 			err: "prometheusQuery discovery requires at least one template",
+		},
+		{
+			config: `rule {
+  name "....+++" {}
+}`,
+			err: "error parsing regexp: invalid nested repetition operator: `++`",
+		},
+		{
+			config: `rule {
+  name "xxx" {
+    severity  = "xxx"
+  }
+}`,
+			err: "unknown severity: xxx",
 		},
 	}